JSRunner: Refactor structure
